Return flow bugs fixes on 26-01-24

+++ src/pages/return/components/single-return/package-details/PackageDetails.js
@@ -46,7 +46,7 @@ export default function PackageDetails({ currentStep, handleChangeStep }) {
         const { id, value, type } = event.target;
         setFormDirectField({
             ...formDirectField,
-            [id]: type === 'number' ? parseInt(value, 10) : value,
+            [id]: type === 'number' ? parseFloat(value, 10) : value,
         });
     };
